added helpfull comments throughout the project
