+++ src/services/repositoryService.ts
@@ -3,13 +3,13 @@ import { Commit } from "../types/Commit";
 import { getAllCommits } from "../fetch/fetchCommits";
 import { getAllFiles } from "../fetch/fetchFiles";
 import { calculateRiskScore } from "./riskScoreService";
-import { FilePath } from "../types/FilePath";
+import { FilePath } from "../types/File";
 import { File } from "../db/models/File";
 import configs from "../configs/fetch.configs.json";
 import { fetchDetailsWithInstallationId } from "../fetch/fetch";
-import { FileType } from "../types/FileType";
+import { FileType } from "../types/File";
 import { TrainingFile } from "../db/models/TrainingFile";
-import { TrainingFileType } from "../types/TrainingFileType";
+import { TrainingFileType } from "../types/File";
 import { retrainPredictorModel } from "./predictionService";
 import { Job } from "../types/Job";
 import { JobModel, JobName, JobStatus } from "../db/models/Job";
