+++ src/services/riskScoreService.ts
@@ -2,6 +2,26 @@ import { Commit } from "../types/Commit";
 import { getTimeDifference } from "../utils";
 import { Probot } from "probot";
 
+/**
+ * Calculates a risk score for a given set of commits based
+ * on their timestamps and distribution.
+ *
+ * The function sorts commits chronologically, calculates the time difference
+ * between each commit and the oldest commit, and assigns a hot spot factor to
+ * each commit based on its temporal distance from the oldest commit.
+ * The risk score is then determined using a logistic function applied to the
+ * accumulated hot spot factors. In case of errors, the function logs warnings
+ * for invalid commit dates and returns a default risk score of 0.
+ *
+ * This function is written by taking inspiration from {@link https://github.com/niedbalski/python-bugspots | bugspots}
+ *
+ * @param {Probot} app The Probot instance.
+ * @param {Commit[]} commits An array of Commit objects representing
+ * the commits to be analyzed.
+ * @returns {number} The calculated risk score, indicating the risk
+ * level associated with the commit history. A higher score suggests a
+ * higher potential risk.
+ */
 export function calculateRiskScore(app: Probot, commits: Commit[]): number {
   try {
     if (commits.length === 0) {
