--- src/types/FilePath.ts
@@ -1,3 +1,0 @@
-export type FilePath = {
-  path: string;
-};
