--- src/types/FileScoreMap.ts
@@ -1,5 +1,0 @@
-export type FileScoreMap = {
-  fileName: string;
-  score: number;
-  predictedScore: number;
-};
