--- src/types/GithubResponseFile.ts
@@ -1,8 +1,0 @@
-import { FileStatus } from "../constants/GithubContants";
-
-export type GithubResponseFile = {
-  sha: string;
-  filePath: string;
-  status: FileStatus;
-  previousFileName?: string;
-};
