--- src/types/TrainingFileType.ts
@@ -1,8 +1,0 @@
-export type TrainingFileType = {
-  installationId: number;
-  owner: string;
-  repoName: string;
-  filePath: string;
-  numberOfCommits: number;
-  riskScore: number;
-};
